[#91][Shaun] updates from CR
